Cast tensors to float in loop operation because stacking of different dtypes is not supported.

+++ code/common/src/compute/expression/evaluate/node.ts
@@ -268,7 +268,7 @@ export const getTypedTensorForLoopOperationNode = (
 
         return {
             tensor: tf.stack(
-                tensorList.map(t => t.tensor),
+                tensorList.map(t => t.tensor.toFloat()),
                 -1
             ),
             isProbabilistic: isAnyProbabilistic,
